+++ studit/src/main/java/com/example/studit/controller/StudyController.java
@@ -1,6 +1,6 @@
 package com.example.studit.controller;
 
-import com.example.studit.config.swagger.BaseResponse;
+import com.example.studit.config.exception.BaseResponse;
 import com.example.studit.domain.enumType.StudyStatus;
 import com.example.studit.domain.study.dto.GetInteriorRes;
 import com.example.studit.domain.study.dto.PatchAddReq;
